fix: tts

+++ apps/web-demo/app/lib/workers/speech-worker-source.js
@@ -40,7 +40,7 @@ try {
   const ttsDevice = hasWebGPU ? "webgpu" : "wasm"; // fallback to wasm if webgpu doens't work
   
   tts = await KokoroTTS.from_pretrained(model_id, {
-    dtype: "fp16",
+    dtype: "fp32",
     device: ttsDevice,
   });
   console.log('TTS initialized successfully');
